+++ src/algorithms/util/standard.h
@@ -34,7 +34,7 @@ void standard_build_tree_from_bits(huffman_node *root, bit_input_stream *input, 
  *
  * @param frequencies the frequency table
  */
-huffman_tree *standard_build_tree_from_frequencies(uint_least32_t frequencies[]);
+huffman_tree *standard_build_tree_from_frequencies(uint_least64_t frequencies[]);
 
 /**
  * Decodes the next character in the input stream.
